% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tabu.R
\name{plot.tabu}
\alias{plot.tabu}
\title{R Based Tabu Search Plot Function}
\usage{
\method{plot}{tabu}(x, type = "default", ...)
}
\arguments{
\item{x}{a tabu object.}

\item{type}{"tracePlot" or default.}

\item{...}{options directly passed to the plot function.}
}
\description{
Plots features of an optimization run of the tabu search algorithm for binary strings. The default plots show: (a) the number of times each element of the string was set to one over the search, (b) frequency of moves for each element of the string over the serach, (c) the number of ones in the chosen configuration at each iteration, (d) the objective function value of the current configuration at each iteration of the algorithm. The "tracePlot" shows the current configurations for all interations.
}
\examples{
# A simple example

evaluateSimple <- function(th)return(1)
result <- tabuSearch(size = 20, iters = 100, objFunc = evaluateSimple)

plot(result)
plot(result, "tracePlot")
}
