% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuSearch-package.R
\docType{package}
\name{tabuSearch-package}
\alias{tabuSearch-package}
\title{tabuSearch: Tabu Search Algorithm for Binary Configurations}
\description{
Tabu search algorithm for binary configurations. A basic version of the algorithm as described by Fouskakis and Draper (2007) \doi{10.1111/j.1751-5823.2002.tb00174.x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/domijan/tabuSearch}
  \item Report bugs at \url{https://github.com/domijan/tabuSearch/issues}
}

}
\author{
\strong{Maintainer}: Katarina Domijan \email{domijank@tcd.ie} (\href{https://orcid.org/0000-0002-4268-2236}{ORCID})

}
\keyword{internal}
