% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tabu.R
\name{summary.tabu}
\alias{summary.tabu}
\title{R Based Tabu Search Summary Function}
\usage{
\method{summary}{tabu}(object, verbose = FALSE, ...)
}
\arguments{
\item{object}{a tabu object.}

\item{verbose}{if true, the optimal configuration(s) will be printed.}

\item{...}{other options (ignored).}
}
\value{
a matrix of configurations from iterations with maximum value of objective function (i.e. Optimum configuration).
}
\description{
Summarizes the results of a tabu search optimization run.
}
\examples{
# A simple example

evaluateSimple <- function(th)return(1)
result <- tabuSearch(size = 20, iters = 100, objFunc = evaluateSimple)

summary(result)
summary(result, verbose = TRUE)

}
